Refactored Layout
