+++ components/util/UnAuthButton.tsx
@@ -1,0 +1,12 @@
+import * as React from 'react'
+import Link from 'next/link'
+
+const UnAuthButton = () => (
+  <div>
+    <Link href="/login">
+      <a className="btn btn-secondary border mr-3">Login</a>
+    </Link>
+  </div>
+)
+
+export default UnAuthButton
